Econometric presentation Presentation
Road fright is one of the major mode of transportation in latvia, followed by rail way. Due to the increase in the transportation, the consumption of fuel also tend to be rising high. (Eng.LSM.lv,Latvian Public Broadcasting).
This study will be an analysis the road freight transport in Latvia by means of volume and its fuel consumption .
The time period of the data is from 2014-2023.

Mathematical and statistical methods based on methods of averages, relative values, calculation and comparison, correlation and grouping pertaining to the gathered data from the mentioned sources have been followed. 